+++ crates/web-sys/src/features/gen_Navigator.rs
@@ -12,6 +12,14 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Navigator`*"]
     pub type Navigator;
+    #[cfg(feature = "Clipboard")]
+    # [wasm_bindgen (structural , method , getter , js_class = "Navigator" , js_name = clipboard)]
+    #[doc = "Getter for the `clipboard` field of this object."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Navigator/clipboard)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Clipboard`, `Navigator`*"]
+    pub fn clipboard(this: &Navigator) -> Clipboard;
     #[cfg(feature = "Permissions")]
     # [wasm_bindgen (structural , catch , method , getter , js_class = "Navigator" , js_name = permissions)]
     #[doc = "Getter for the `permissions` field of this object."]
@@ -125,18 +133,6 @@ extern "C" {
     #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
     #[doc = "[described in the `wasm-bindgen` guide](https://rustwasm.github.io/docs/wasm-bindgen/web-sys/unstable-apis.html)*"]
     pub fn bluetooth(this: &Navigator) -> Option<Bluetooth>;
-    #[cfg(web_sys_unstable_apis)]
-    #[cfg(feature = "Clipboard")]
-    # [wasm_bindgen (structural , method , getter , js_class = "Navigator" , js_name = clipboard)]
-    #[doc = "Getter for the `clipboard` field of this object."]
-    #[doc = ""]
-    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Navigator/clipboard)"]
-    #[doc = ""]
-    #[doc = "*This API requires the following crate features to be activated: `Clipboard`, `Navigator`*"]
-    #[doc = ""]
-    #[doc = "*This API is unstable and requires `--cfg=web_sys_unstable_apis` to be activated, as"]
-    #[doc = "[described in the `wasm-bindgen` guide](https://rustwasm.github.io/docs/wasm-bindgen/web-sys/unstable-apis.html)*"]
-    pub fn clipboard(this: &Navigator) -> Option<Clipboard>;
     #[cfg(web_sys_unstable_apis)]
     #[cfg(feature = "MediaSession")]
     # [wasm_bindgen (structural , method , getter , js_class = "Navigator" , js_name = mediaSession)]